--- prover/rust-toolchain
@@ -1,1 +1,0 @@
-nightly-2023-12-03
